+++ CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.4.3]
+## [1.5.0]
+
+### Added
+- Course addition search support for names (hebrew and english).
+- Dropdown menu for course addition search results.
+- Placeholder text for username, password and ID inputs in options page.
 
 ### Changed
 - Updated icons folder names for better clarity.